backend/requirements.txt
Flask
pymongo==4.6.3
python-dotenv==0.21.0
flask-cors
flask-sqlalchemy